FIX server.js to customize it in other projects
