+++ webpack.server.config.js
@@ -25,6 +25,9 @@ const voltranServerConfigPath = voltranConfig.webpackConfiguration.server;
 const voltranServerConfig = voltranServerConfigPath
   ? require(voltranConfig.webpackConfiguration.server)
   : '';
+const voltranServer = voltranConfig.entry
+  ? voltranConfig.entry
+  : path.resolve(__dirname, isDebug ? 'src/server.js' : 'src/main.js');
 
 const serverConfig = webpackMerge(commonConfig, voltranServerConfig, {
   name: 'server',
@@ -34,7 +37,7 @@ const serverConfig = webpackMerge(commonConfig, voltranServerConfig, {
   mode: isDebug ? 'development' : 'production',
 
   entry: {
-    server: [path.resolve(__dirname, isDebug ? 'src/server.js' : 'src/main.js')]
+    server: [voltranServer]
   },
 
   output: {
